extfs: Add an inode->vnode table to fix hardlink bugs
Previously, ext_mount would always create a new vnode for each dirent, regardless of if there is already a vnode for that inode. This commit fixes that by introducing a inode->vnode table in the ext cache.
This also solves a bug with mounting the same partition in two different spots